Refectored page objkect to be easier to use and maintain. Added new test for editing workitem.
