+++ src/tests/work-item/work-item-list/work-item-list.page.js
@@ -16,6 +16,10 @@ var WorkItemListPage = function () {
   browser.get("http://localhost:8088/?token=justarandomtokenfortest");
 };
 
+var testSupport = require('./testSupport'),
+  WorkItemDetailPage = require('./work-item-detail.page'),
+  WorkItemBoardPage = require('./work-item-board.page');
+
 var until = protractor.ExpectedConditions;
 var waitTime = 30000;
 
@@ -307,198 +311,6 @@ workItemByTitle:  {
       return button.click(); }
   },
 
-});
-
-
-/*
- * Work Item Detail Page Definition
- */
-
-var WorkItemDetailPage = function (idValue) {
-  browser.get("http://localhost:8088/detail/" + idValue + "?token=justarandomtokenfortest");
-};
-
-WorkItemDetailPage.prototype  = Object.create({}, {
-
-  workItemDetailPageTitle:  {
-    get: function ()
-    { return element(by.xpath(".//*[@id='workItemDetail_Wrapper']/h1")); }
-  },
-
-  workItemDetailTitle:  {
-    get: function ()
-    { return element(by.id("wi-detail-title")); }
-  },
-
-  setWorkItemDetailTitle:  {
-    value: function (newTitleString, append)
-    {   if (!append) {this.workItemDetailTitle.clear(newTitleString)};
-    return this.workItemDetailTitle.sendKeys(newTitleString); }
-  },
-
-  workItemDetailDescription:  {
-    get: function ()
-    { return element(by.id("wi-detail-desc")); }
-  },
-
-  setWorkItemDetailDescription:  {
-    value: function (newDescriptionString, append)
-    {   if (!append) {this.workItemDetailDescription.clear(newDescriptionString)};
-    return this.workItemDetailDescription.sendKeys(newDescriptionString); }
-  },
-
-  workItemDetailType:  {
-    get: function ()
-    { return element(by.id("wi-detail-type")); }
-  },
-
-  setWorkItemDetailType:  {
-    value: function (newTypeString, append)
-    {   return this.workItemDetailType.sendKeys(newTypeString); }
-  },
-
-  workItemDetailCreator:  {
-    get: function ()
-    { return element(by.id("wi-detail-creator")); }
-  },
-
-  setWorkItemDetailCreator:  {
-    value: function (newCreatorString, append)
-    {   if (!append) {this.workItemDetailCreator.clear(newCreatorString)};
-    return this.workItemDetailCreator.sendKeys(newCreatorString); }
-  },
-
-  workItemDetailAssignee:  {
-    get: function ()
-    { return element(by.id("wi-detail-assignee")); }
-  },
-
-  setWorkItemDetailAssignee:  {
-    value: function (newAssigneeString, append)
-    {   if (!append) {this.workItemDetailAssignee.clear(newAssigneeString)};
-    return this.workItemDetailAssignee.sendKeys(newAssigneeString); }
-  },
-
-  workItemDetailState:  {
-    get: function ()
-    { return element(by.id("wi-detail-state")); }
-  },
-
-  setWorkItemDetailState:  {
-    value: function (newStateString, append)
-    {   return this.workItemDetailState.sendKeys(newStateString); }
-  },
-
-  workItemDetailSaveButton:  {
-    get: function ()
-    { return element(by.css(".btn.btn-primary")); }
-  },
-
-  clickWorkItemDetailSaveButton:   {
-    value: function ()
-    { 
-      browser.wait(until.presenceOf(this.workItemDetailSaveButton), waitTime, 'Failed to find the workItemDetailSaveButton');
-      return this.workItemDetailSaveButton.click(); }
-  },
-
-  workItemDetailCancelButton:  {
-    get: function ()
-    { return element(by.css(".btn.btn-default")); }
-  },
-
-  clickWorkItemDetailCancelButton:   {
-    value: function ()
-    { 
-      browser.wait(until.presenceOf(this.workItemDetailCancelButton), waitTime, 'Failed to find the workItemDetailCancelButton');
-      return this.workItemDetailCancelButton.click(); }
-  }
-
-});
-
-/*
- * Work Item Board Page Definition
- */
-
-var WorkItemBoardPage = function () {
-  browser.get("http://localhost:8088/#/board");
-};
-
-WorkItemBoardPage.prototype  = Object.create({}, {
-
-  workItemListButton:  {
-    get: function ()
-    { return element(by.id("header_menuWorkItemList")); }
-  },
-
-  boardButton:  {
-    get: function ()
-    { return element(by.id("header_menuBoard")); }
-  },
-
-  workItemBoardSearchBox:  {
-    get: function ()
-    { return element(by.css("#search-box input")); }
-  },
-
-  typeworkItemBoardSearchBox:  {
-    value: function (keys)
-    { return this.workItemBoardSearchBox.sendKeys(keys); }
-  },
-
-  allWorkItemCards:  {
-    get: function ()
-    { return element.all(by.css("#board_topWorkItems")); }
-  },
-
-  firstWorkItem:  {
-    get: function ()
-    { return element.all(by.css("#board_topWorkItems")).first(); }
-  },
-
-  lastWorkItem:  {
-    get: function ()
-    { return element.all(by.css("#board_topWorkItems")).last(); }
-  },
-
-  workItemByIndex:  {
-    value: function (itemNumber)
-      { return element.all(by.css("#board_topWorkItems")).get(itemNumber); }
-    },
-
-  workItemByNumber:  {
-    value: function (itemNumber)
-    {
-      var xPathString = "board_topWorkItemList_" + itemNumber;
-      return element(by.id(xPathString));
-    }
-  },
-
-  /*
-   * When the Work Item 'View Detail' page is opened, there can be a delay of a few seconds before
-   * the page contents are displayed - the browser.wait statement covers this wait for the title
-   * of the page - there is a further delay before the values of the elements on the page are displayed.
-   */
-  clickWorkItemViewButton:   {
-    value: function (button, idValue)
-    {
-      button.click();
-      var theDetailPage = new WorkItemDetailPage (idValue);
-      var until = protractor.ExpectedConditions;
-      browser.wait(until.presenceOf(theDetailPage.workItemDetailPageTitle), waitTime, 'Detail page title taking too long to appear in the DOM');
-      browser.wait(waitForText(theDetailPage.workItemDetailTitle), waitTime, "Title text is still not present");
-      return theDetailPage;
-    }
-  },
-
-  userToggle:  {
-    get: function ()
-    { return element(by.id("header_dropdownToggle")); }
-  },
-
-  clickUserToggle:   {
-    value: function ()
-    { return this.userToggle.click(); }
-  }
 });
 
 /*
