+++ src/tests/work-item/work-item-list/work-item-list.spec.js
@@ -13,7 +13,8 @@
  */
 
 var WorkItemListPage = require('./work-item-list.page'),
-  testSupport = require('./testSupport');
+  testSupport = require('./testSupport'),
+  WorkItemDetailPage = require('./work-item-detail.page');
 
 describe('Work item list', function () {
   var page, items, startCount, browserMode;
@@ -72,7 +73,7 @@ describe('Work item list', function () {
   it('should contain right mock data on detail page - desktop.', function() { 
     testSupport.setBrowserMode('desktop');	
     page.workItemViewId(page.firstWorkItem).getText().then(function (text) { 
-      detailPage = page.clickWorkItemViewButton(page.workItemViewButton(page.firstWorkItem), text);
+      var detailPage = page.clickWorkItemViewButton(page.workItemViewButton(page.firstWorkItem), text);
       expect(detailPage.workItemDetailPageTitle.getText()).toBe(workItemMockData.pageTitle);
       expect(detailPage.workItemDetailTitle.getAttribute("value")).toBe(workItemMockData.workItemTitle);      
       expect(detailPage.workItemDetailDescription.getAttribute("value")).toBe(workItemMockData.workItemDescription);
